refactor(foundation)!: split base.py into dedicated error modules
