+++ src/forging_blocks/foundation/errors/cant_modify_immutable_attribute_error.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from forging_blocks.foundation.errors.base import Error
+from forging_blocks.foundation.errors.error import Error
 from forging_blocks.foundation.errors.core import ErrorMessage, ErrorMetadata
