+++ src/forging_blocks/foundation/errors/none_not_allowed_error.py
@@ -1,0 +1,7 @@
+"""Error indicating that a None value was provided where it is not allowed."""
+
+from forging_blocks.foundation.errors.error import Error
+
+
+class NoneNotAllowedError(Error):
+    """Error indicating that a None value was provided where it is not allowed."""
