+++ src/forging_blocks/foundation/errors/result_access_error.py
@@ -1,5 +1,5 @@
 """Errors raised when accessing Result.value or Result.error on the wrong variant"""
-from forging_blocks.foundation.errors.base import Error
+from forging_blocks.foundation.errors.error import Error
 from forging_blocks.foundation.errors.core import ErrorMessage
